+++ app/types/common/reviews.types.ts
@@ -1,26 +1,27 @@
-export type ReviewListType = {
-  data: {
+export type ReviewListDataType = {
+  teamId: number;
+  id: number;
+  score: number;
+  comment: string;
+  createdAt: string;
+  Gathering: {
     teamId: number;
     id: number;
-    score: number;
-    comment: string;
-    createdAt: string;
-    Gathering: {
-      teamId: number;
-      id: number;
-      type: string;
-      name: string;
-      dateTime: string;
-      location: string;
-      image: string;
-    };
-    User: {
-      teamId: number;
-      id: number;
-      name: string;
-      image: string;
-    };
-  }[];
+    type: string;
+    name: string;
+    dateTime: string;
+    location: string;
+    image: string;
+  };
+  User: {
+    teamId: number;
+    id: number;
+    name: string;
+    image: string;
+  };
+};
+export type ReviewListType = {
+  data: ReviewListDataType[];
   totalItemCount: number;
   currentPage: number;
   totalPages: number;
